requirements.txt
Django==3.0.8
gunicorn==20.0.4
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
